Sans platines

+++ src/trameimagecameroa.h
@@ -29,11 +29,11 @@ struct EnteteTrameDonneesCamerOA
 	char chaine_magique[TAILLE_CH_MAG_ENTDONCAMEROA];	// Chaine magique empreinte pour reconnaitre le debut de la trame video
 	char type_trame;									// Le type de trame image
 	char type_image;									// Le type d'image en fonction du type de trame
-	unsigned long tx;									// La dimension en abscisse de l'image  (on va envoyer de l'unsigned short 16 bits)
-	unsigned long ty;									// La dimension en ordonnee de l'image
+	uint32_t tx;									// La dimension en abscisse de l'image  (on va envoyer de l'unsigned short 16 bits)
+	uint32_t ty;									// La dimension en ordonnee de l'image
 	QDateTime debut;									// Heure du debut de pose
 	QDateTime fin;										// Heure de fin de pose
-	int duree;											// Duree de la pose en 1/10000 de seconde
+	int32_t duree;											// Duree de la pose en 1/10000 de seconde
 };
 
 #endif
